reportReview 컨트롤러에서 값 set하던거 처리했음
