+++ src/main/java/shop/mtcoding/finalproject/web/OrderApiController.java
@@ -29,27 +29,29 @@ public class OrderApiController {
     private final Logger log = LoggerFactory.getLogger(getClass());
     private final OrderService orderService;
 
-    @GetMapping("/order/history/list")
-    public ResponseEntity<?> getOrderHistoryList(@AuthenticationPrincipal LoginUser loginUser) {
+    @GetMapping("/user/{userId}/order/history/list")
+    public ResponseEntity<?> getOrderHistoryList(@PathVariable Long userId,
+            @AuthenticationPrincipal LoginUser loginUser) {
+        loginUser.getUser().checkAccount(userId);
         log.debug("디버그 : 컨트롤러 응답 전");
-        OrderHistoryListRespDto orderHistoryListRespDto = orderService.주문내역_목록보기(loginUser.getUser().getId());
+        OrderHistoryListRespDto orderHistoryListRespDto = orderService.주문내역_목록보기(userId);
         return new ResponseEntity<>(new ResponseDto<>("주문내역 목록보기 성공", orderHistoryListRespDto), HttpStatus.OK);
     }
 
     /* 승현 작업 시작 */
 
-    @PutMapping("/store/{storeId}/order/{orderId}/state")
+    @PutMapping("/user/{userId}/store/{storeId}/order/{orderId}/state")
     public ResponseEntity<?> UpdateOrderByUserIdToComplete(@PathVariable Long storeId, @PathVariable Long orderId,
-            @RequestBody UpdateToCancleOrderReqDto updateToCancleOrderReqDto,
+            @RequestBody UpdateToCancleOrderReqDto updateToCancleOrderReqDto, @PathVariable Long userId,
             @AuthenticationPrincipal LoginUser loginUser) {
-        updateToCancleOrderReqDto.setUserId(loginUser.getUser().getId());
+        updateToCancleOrderReqDto.setUserId(userId);
         updateToCancleOrderReqDto.setOrderId(orderId);
         updateToCancleOrderReqDto.setStoreId(storeId);
         String state = orderService.updatToState(updateToCancleOrderReqDto);
         return new ResponseEntity<>(new ResponseDto<>("주문상태 변경완료", state), HttpStatus.OK);
     }
 
-    @GetMapping("/store/{storeId}/order")
+    @GetMapping("/user/{userId}/store/{storeId}/order")
     public ResponseEntity<?> findAllByStoreId(@PathVariable Long storeId,
             @AuthenticationPrincipal LoginUser loginUser) {
         List<ShowOrderListRespDto> showOrderListRespDtoList = orderService.findAllByStoreId(storeId,
